fixed rust file size exceeded
